Move xml writer code to separate module
Convenience method tries to auto select correct
writer based on output_path.

S3 upload methods use in memory streams to avoid
the extra overhead of writing gzipped data to a
file just to upload and then delete it.

Moves code in onramp/__init__.py to onramp/diff.py,
now __init__.py correctly exposes the public API
